Provide data directly to DownloadFile
This removes the need for the DownloadSessionData component and makes
the DownloadFile component more versatile.
