+++ src/components/DownloadErrorData.js
@@ -21,13 +21,10 @@ class DownloadErrorData extends React.PureComponent {
         stack: error.stack
       }
     };
-    const json = JSON.stringify(data);
-    const blob = new Blob([json], { type: 'application/json' });
-    const url = URL.createObjectURL(blob);
     const filename = `cliff-effects-${new Date().toISOString().slice(0, 10)}`
   
     return (
-      <DownloadFileAs url={url} defaultFilename={filename}>
+      <DownloadFileAs data={data} defaultFilename={filename}>
         Download error report
       </DownloadFileAs>
     );
