+++ src/components/DownloadFile.js
@@ -5,12 +5,19 @@ import React from 'react';
  * 
  * @param props {object}
  * @param props.children {string|null} - Optional link text
+ * @param props.data {object} - Data to be saved
  * @param props.filename {string}
- * @param props.url {string}
  */
-const DownloadFile = ({ children = "Download file", filename, url }) =>
-  <a href={url} download={filename}>
-    {children}
-  </a>;
+const DownloadFile = ({ children = 'Download file', data, filename = 'cliff-effects.json' }) => {
+  const json = JSON.stringify(data);
+  const blob = new Blob([json], { type: 'application/json' });
+  const url = URL.createObjectURL(blob);
+  
+  return (
+    <a href={url} download={filename}>
+      {children}
+    </a>
+  );
+};
 
 export default DownloadFile;
